pyproject.toml
[project]
name = "planloop"
version = "0.1.0"
dependencies = [
    "typer",
    "rich",
    "python-slugify",
    "openai",
]

[project.scripts]
planloop = "planloop.cli:app"